improve component reusability
